Bug 906176: Move to using Django staticfiles for media.
This keeps the URL for CSS/JS/Images/Etc. as /media/ so that
all of the .less files and urls around the net need not change.
The new MEDIA_URL is '/user_media/', and the STATIC_URL is '/media/'.
The files all stay in the 'media' folder in the project, but they'll now
be served from the 'static' folder after 'collectstatic' is run.

All templates, views, and tests have been updated for the new changes,
as well as the l10n media helpers. The site works well for me in
spot-checks, and the tests pass.

One caveat is that you must run 'manage.py collectstatic' before
running tests due to the way jingo-minify works regarding the
compilation of .less files when not in DEBUG mode. I've yet to find
a way around this.