Bug 906176: Enable caching static storage and remove cachebusts.
Removes all of the "?2015-01" mess from images around the site.
Enables the built in django processor that will add the md5 hash
of the file contents to the file name then use that name in the
html when we use the "static()" helper in templates. The post-processor
also handles references to images in the CSS.