Reorder deployment to keep the git repo clean.
Deployment updates the code via git, then syncs
the files to another directory for processing,
then syncs the files to the servers. It was doing
a lot of processing in the git repo thus creating
a lot of extra files that can cause problems. Moving
this creation to the destination dir means rsync will
clean these up every time and the git repo remains
pristine.

* Don't die if there is no media/revision.txt.
* Allow running of some commands from src dir.